Broke solver_functors.hpp into several different files
