+++ src/restruct_poc/beams_input.cpp
@@ -8,88 +8,9 @@
 #include "InterpolateQPState.hpp"
 #include "InterpolateQPVelocity.hpp"
 #include "InterpolateQPAcceleration.hpp"
+#include "Interpolation.hpp"
 
 namespace openturbine {
-
-void LinearInterpWeights(
-    const double x, const std::vector<double>& xs, std::vector<double>& weights
-) {
-    // Get number of nodes
-    auto n = xs.size();
-
-    // Resize weights and fill with zeros
-    weights.clear();
-    weights.resize(n, 0.);
-
-    // Find the lower bound location
-    auto lower = std::lower_bound(xs.begin(), xs.end(), x);
-
-    // If x is less than the first location
-    if (lower == xs.begin()) {
-        weights.front() = 1.0;
-    }
-
-    // If x is greater than the last location
-    else if (lower == xs.end()) {
-        weights.back() = 1.0;
-    }
-
-    // x is between two points in xs
-    else {
-        size_t index = lower - xs.begin();
-        double lower_loc = xs[index - 1];
-        double upper_loc = xs[index];
-        double weight_upper = (x - lower_loc) / (upper_loc - lower_loc);
-        weights[index - 1] = 1.0 - weight_upper;
-        weights[index] = weight_upper;
-    }
-}
-
-void LagrangePolynomialInterpWeights(
-    const double x, const std::vector<double>& xs, std::vector<double>& weights
-) {
-    // Get number of nodes
-    const auto n = xs.size();
-
-    // Resize weights and fill with 1s
-    weights.clear();
-    weights.resize(n, 1.);
-
-    // Calculate weights
-    for (size_t j = 0; j < n; ++j) {
-        for (size_t m = 0; m < n; ++m) {
-            if (j != m) {
-                weights[j] *= (x - xs[m]) / (xs[j] - xs[m]);
-            }
-        }
-    }
-}
-
-void LagrangePolynomialDerivWeights(
-    const double x, const std::vector<double>& xs, std::vector<double>& weights
-) {
-    // Get number of nodes
-    const auto n = xs.size();
-
-    // Resize weights and fill with zeros
-    weights.clear();
-    weights.resize(n, 0.);
-
-    // Calculate weights
-    for (size_t i = 0; i < n; ++i) {
-        for (size_t j = 0; j < n; ++j) {
-            if (j != i) {
-                double prod = 1.0;
-                for (size_t k = 0; k < n; ++k) {
-                    if (k != i && k != j) {
-                        prod *= (x - xs[k]) / (xs[i] - xs[k]);
-                    }
-                }
-                weights[i] += prod / (xs[i] - xs[j]);
-            }
-        }
-    }
-}
 
 template <typename T1, typename T2, typename T3, typename T4, typename T5, typename T6>
 void PopulateElementViews(
